Insumos

+++ db/migrate/20180125134518_add_peso_to_perro.rb
@@ -1,5 +1,5 @@
 class AddPesoToPerro < ActiveRecord::Migration[5.1]
   def change
-    add_column :perros, :peso, :bigint
+    add_column :perros, :peso, :numeric
   end
 end
